fix(cross-platform): condicionalizar includes Windows-only (dinput.h, d3dx8math.h, atlbase.h) + adicionar typedef Byte para compat com Upstream
- Sincronizei GeneralsMD/Code/GameEngine com upstream main
- Condicionalizado WIN32 defines em PreRTS.h para atlbase.h, windows.h
- Condicionalizado DirectInput em PreRTS.h e KeyDefs.h com #ifdef _WIN32
- Copiado d3dx8_vulkan_math_compat.h para GeneralsMD/Code/GameEngine/Include/Common
- Adicionado Byte typedef em GeneralsMD BaseType.h como alias para UnsignedByte
- Compilação progrediu para 108/1129 objetos

+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/PopupHostGame.cpp
@@ -412,7 +412,7 @@ WindowMsgHandledType PopupHostGameInput( GameWindow *window, UnsignedInt msg, Wi
 					if( BitIsSet( state, KEY_STATE_UP ) )
 					{
 						TheWindowManager->winSendSystemMsg( window, GBM_SELECTED,
-																							(WindowMsgData)(uintptr_t)buttonCancel, buttonCancelID );
+																							(WindowMsgData)buttonCancel, buttonCancelID );
 
 					}
 
@@ -508,7 +508,7 @@ WindowMsgHandledType PopupHostGameSystem( GameWindow *window, UnsignedInt msg, W
 				{
 					if (pos >= 0)
 					{
-						Int ladderID = (Int)(uintptr_t)GadgetComboBoxGetItemData(control, pos);
+						Int ladderID = (Int)GadgetComboBoxGetItemData(control, pos);
 						if (ladderID < 0)
 						{
 							// "Choose a ladder" selected - open overlay
@@ -597,7 +597,7 @@ void createGame( void )
 	req.stagingRoomCreation.ladPort = 0;
 	if (ladderSelectPos >= 0)
 	{
-		ladderID = (Int)(uintptr_t)GadgetComboBoxGetItemData(comboBoxLadderName, ladderSelectPos);
+		ladderID = (Int)GadgetComboBoxGetItemData(comboBoxLadderName, ladderSelectPos);
 		if (ladderID != 0)
 		{
 			// actual ladder
